+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 * `Tag` element got new property for `backgroundColor`
 * Styles for footnotes and task lists, rendered by Markdown GFM parser.
 * React-Flow `NodeContent` element can now be extented by `contentExtension` property containing a `NodeContentExtension` element.
+* `ReactFlow` element with `configuration` property to load it with pre-set configurations for node and edge types.
 
 ### Changed
 * `TextField` elements are using `fullWidth=true` by default